+++ src/rcpp_respiration.cpp
@@ -52,7 +52,7 @@ void rcpp_respiration(Rcpp::NumericMatrix fishpop,
     // Oxycaloric coefficient in J/gO2 consumed multiplied by the energy-density of fish
     // to result in unit of tick^-1
     double respiration = (resp_intercept * std::pow(fishpop(i, 6), resp_slope) *
-                          temp_dependence * fishpop(i, 9)) * 13560.0 * (1.0 / 4800.0);
+                          temp_dependence * fishpop(i, 7)) * 13560.0 * (1.0 / 4800.0);
 
     // check if finite number
     bool check_finite = std::isfinite(respiration);
@@ -61,13 +61,13 @@ void rcpp_respiration(Rcpp::NumericMatrix fishpop,
     if (check_finite) {
 
       // update respiration col
-      fishpop(i, 10) = respiration;
+      fishpop(i, 8) = respiration;
 
     // respiration is infinite (divided by zero probably), use 1 instead
     } else {
 
       // update respiration col
-      fishpop(i, 10) = 1;
+      fishpop(i, 8) = 1;
 
     }
   }