+++ src/rcpp_seagrass_growth.cpp
@@ -32,8 +32,7 @@ void rcpp_seagrass_growth(Rcpp::NumericMatrix seafloor, Rcpp::NumericVector cell
                           double bg_biomass_max, double bg_biomass_min,
                           double ag_biomass_max, double ag_biomass_min,
                           double seagrass_thres, double seagrass_slope,
-                          double seagrass_slough,
-                          double time_frac) {
+                          double seagrass_slough, double time_frac) {
 
   // loop through all seafloor cells
   for (int i = 0; i < seafloor.nrow(); i++) {