+++ src/rcpp_seagrass_growth.h
@@ -11,7 +11,6 @@ void rcpp_seagrass_growth(Rcpp::NumericMatrix seafloor, Rcpp::NumericVector cell
                           double bg_biomass_max, double bg_biomass_min,
                           double ag_biomass_max, double ag_biomass_min,
                           double seagrass_thres, double seagrass_slope,
-                          double seagrass_slough,
-                          double time_frac);
+                          double seagrass_slough, double time_frac);
 
 //#endif // RCPP_SEAGRASS_GROWTH